move configinfo to its own file
